Leverage partial class to implement interface and fix those #region stylecop warnings.
